- fix bug where cropping values of past year were used - fix "cannot read property deleteAll" bug in maps component - increase replication speed by enabling an offline user database per user, removing the need to delete the old database

+++ components/maps.vue
@@ -171,7 +171,7 @@ export default {
       this.Draw.delete(data.features[0].id)
     },
     removeDraw() {
-      this.Draw.deleteAll()
+      if (this.Draw) this.Draw.deleteAll()
     },
     select(data) {
       console.log(data)
